Update to script.js
updated the scoring system to better match a tier list based on "who wants to be a millionaire" and created a function with an if else if statement comparing the number of questions answered correctly to display the proper score.
Also added a temporary session high score that is reset when the page is refreshed so I wouldn't have to use the localStorage funciton.